Removed dependency of the telescope and source objects inside the diffractive measurement strategy to make the process agnostic of the other objects in preparation for further parallelization.
Removed multiple objects (OPDs) support at the wfs_measure parameters, this will be handle independently by the LightPath class, which is easier for the user to setup, maintein and parallelize.

Removed spatial filter check inside the WFS module, this is optically applied to the light BEFORE arriving to the WFS, so the LightPath class will add it at the right optical location to improve transparency and mantenabiity.

Input phase is now received as a parameter, not consulted globally to avoid problems at a later stage when parallelization is done.