Modified calibration procedure inside initialize_wfs to avoid the use of a class attribute for the telescope/src. They are now provided as parameters.
Removed dependencies on the class attribute telescope EXCEPT for the geometric WFS that was not addressed (for now). The objective is to provide the phase and telescopes/sources features directly to the methods to avoid problems during parallelization AND to improve the maintenability, the methods will only work with parameters provided and with physical properties of the WFS instead of reading indirectly the porperties of external objects, which might go wrong if it is not well controlled.