Calibration procedure is modified to generate locally the null phase instead of using the telescope and source properties.
Camera id modified to be called directly instead of using the operator *.

Raw(ideal) frame, signal and signal_2D are returned after WFS integration. They are now local variables instead of class attributes, enabling parallel processing using multiple sources without requiring multiple WFS. Hence, the WFS shall be change ONLY if its physical properties vary.

TODO: address no pupil phase for the geometric strategy -> it is required currently due to the gradient computation.

IntegrationTime was added to the _integrated_time property of the Detector class, but it was not defined in the WFS constructor when the self.cam object is instantiated, so the Detector was ignoring this parameter. For the WFS it does not make sense to have an integration time in the camera different from the loop cycle, so I removed it completely